Move PixMapHelper class into its own file
